feat(checkbox): ✨ finalize the new rfc
