+++ src/checkbox/CheckboxInput.tsx
@@ -1,49 +1,43 @@
 import {
-  Checkbox as ReakitCheckbox,
-  CheckboxProps as ReakitCheckboxProps,
-} from "reakit";
+  useCheckbox,
+  CheckboxOptions,
+  CheckboxHTMLProps,
+} from "../checkboxReakit";
 import { cx } from "@renderlesskit/react";
+import { createComponent, createHook } from "reakit-system";
 
 import { useTheme } from "../theme";
-import { useCheckboxContext } from "./Checkbox";
-import { forwardRefWithAs } from "../utils/types";
-import { useFormControl, CommonFieldProps } from "../form-field";
+import { CHECKBOX_INPUT_KEYS } from "./__keys";
+import { CheckboxStateReturn } from "./CheckboxState";
 
-export type CheckboxInputProps = ReakitCheckboxProps &
-  Omit<CommonFieldProps, "id" | "isReadOnly">;
+export type CheckboxInputOptions = CheckboxOptions & {
+  size: CheckboxStateReturn["size"];
+};
 
-export const CheckboxInput = forwardRefWithAs<
-  CheckboxInputProps,
-  HTMLInputElement,
-  "input"
->((props, ref) => {
-  const { className, ...rest } = props;
-  const {
-    state: { isInvalid, ...state },
-  } = useCheckboxContext();
-  // Interpts with the checked unchecked state
-  if (state["value"] === undefined) delete state["value"];
-  if (state["checked"] === undefined) delete state["checked"];
+export type CheckboxInputHTMLProps = Omit<CheckboxHTMLProps, "size">;
 
-  const theme = useTheme();
-  const checkboxInputStyles = cx(theme.checkbox.input, className);
+export type CheckboxInputProps = CheckboxInputOptions & CheckboxInputHTMLProps;
 
-  const formFieldProps = useFormControl({
-    ...rest,
-    isDisabled: state.disabled || props.isDisabled || props.disabled,
-    isInvalid:
-      isInvalid || props.isInvalid || (props["aria-invalid"] as boolean),
-    isRequired: props.isRequired || props.required,
-  });
+export const useCheckboxInput = createHook<
+  CheckboxInputOptions,
+  CheckboxInputHTMLProps
+>({
+  name: "CheckboxInput",
+  compose: useCheckbox,
+  keys: CHECKBOX_INPUT_KEYS,
 
-  return (
-    <ReakitCheckbox
-      ref={ref}
-      className={checkboxInputStyles}
-      {...state}
-      {...formFieldProps}
-    />
-  );
+  useProps(options, htmlProps) {
+    const { className: htmlClassName, ...restHtmlProps } = htmlProps;
+
+    const theme = useTheme("checkbox");
+    const className = cx(theme.input, htmlClassName);
+
+    return { className, ...restHtmlProps };
+  },
 });
 
-CheckboxInput.displayName = "CheckboxInput";
+export const CheckboxInput = createComponent({
+  as: "input",
+  memo: true,
+  useHook: useCheckboxInput,
+});
