+++ src/checkbox/index.ts
@@ -1,6 +1,7 @@
 export * from "./Checkbox";
+export * from "./CheckboxState";
 export * from "./CheckboxIcon";
-export * from "./CheckboxInput";
-export * from "./CheckboxLabel";
 export * from "./CheckboxText";
 export * from "./CheckboxDescription";
+export * from "./CheckboxLabel";
+export * from "./CheckboxInput";
