--- src/checkboxNew/helpers.tsx
@@ -1,24 +1,0 @@
-import { Dict } from "../utils/types";
-import { getValidChildren } from "../utils";
-
-const ComponentPropsMap = {
-  CheckboxLabel: "labelProps",
-  CheckboxInput: "inputProps",
-  CheckboxIcon: "iconProps",
-  CheckboxText: "textProps",
-  CheckboxDescription: "descriptionProps",
-};
-
-export const getCheckboxComponentProps = (children: React.ReactNode) => {
-  const validChildren = getValidChildren(children);
-  const props: Dict = {};
-
-  validChildren.forEach(child => {
-    props[
-      // @ts-ignore
-      ComponentPropsMap[child.type.displayName]
-    ] = child.props;
-  });
-
-  return props;
-};
