--- src/checkboxNew/index.ts
@@ -1,7 +1,0 @@
-export * from "./Checkbox";
-export * from "./CheckboxState";
-export * from "./CheckboxIcon";
-export * from "./CheckboxText";
-export * from "./CheckboxDescription";
-export * from "./CheckboxLabel";
-export * from "./CheckboxInput";
