+++ src/theme/defaultTheme/checkbox.ts
@@ -1,58 +1,62 @@
-const field = {
-  base: "lib:flex lib:flex-col",
-  description: {
-    base: "lib:text-gray-500 lib:mt-1 lib:text-sm",
+export const checkbox = {
+  label:
+    "lib:relative lib:inline-flex lib:items-center lib:justify-center lib:cursor-pointer lib:align-top",
+  input: "peer lib:sr-only",
+  icon: {
+    base: "lib:inline-flex lib:items-center lib:justify-center lib:self-start lib:flex-shrink-0 lib:align-top lib:select-none lib:transition-all",
     size: {
-      sm: "lib:text-xs",
-      md: "lib:text-sm",
-      lg: "lib:text-sm",
+      sm: "lib:w-3.5 lib:h-3.5 lib:text-[10px] lib:border-[1.5px] lib:rounded",
+      md: "lib:w-4 lib:h-4 lib:text-xs lib:border-[1.5px] lib:rounded",
+      lg: "lib:w-5 lib:h-5 lib:text-base lib:border-[1.5px] lib:rounded",
+    },
+    checked: {
+      default: "lib:bg-gray-800 lib:border-gray-800 lib:text-white",
+      hover: "peer-hover:lib:bg-gray-700 peer-hover:lib:border-gray-700",
+      active: "peer-active:lib:bg-gray-800 peer-active:lib:border-gray-800",
+      focus:
+        "peer-focus-visible:lib:ring-2 peer-focus-visible:lib:ring-gray-300",
+      disabled:
+        "peer-disabled:lib:bg-gray-200 peer-disabled:lib:border-gray-200 peer-disabled:lib:text-gray-400",
+      invalid:
+        "lib:bg-red-600 lib:border-red-600 lib:text-white peer-hover:lib:bg-red-500 peer-hover:lib:border-red-500 peer-active:lib:bg-red-600 peer-active:lib:border-red-600 peer-focus-visible:lib:ring-2 peer-focus-visible:lib:ring-red-300 peer-disabled:lib:bg-red-200 peer-disabled:lib:border-red-200 peer-disabled:lib:text-red-400",
+    },
+    unChecked: {
+      default: "lib:bg-white lib:border-gray-400 lib:text-current",
+      hover: "peer-hover:lib:border-gray-500",
+      active: "peer-active:lib:border-gray-400",
+      focus:
+        "peer-focus-visible:lib:ring-2 peer-focus-visible:lib:ring-gray-300",
+      disabled:
+        "peer-disabled:lib:bg-gray-200 peer-disabled:lib:border-gray-300",
+      invalid:
+        "lib:bg-white lib:border-red-600 lib:text-current peer-hover:lib:border-red-700 peer-active:lib:border-red-600 peer-focus-visible:lib:ring-2 peer-focus-visible:lib:ring-red-300 peer-disabled:lib:bg-red-200 peer-disabled:lib:border-red-300",
+    },
+    indeterminate: {
+      default: "lib:bg-gray-800 lib:border-gray-800 lib:text-white",
+      hover: "peer-hover:lib:bg-gray-700 peer-hover:lib:border-gray-700",
+      active: "peer-active:lib:bg-gray-800 peer-active:lib:border-gray-800",
+      focus:
+        "peer-focus-visible:lib:ring-2 peer-focus-visible:lib:ring-gray-300",
+      disabled:
+        "peer-disabled:lib:bg-gray-200 peer-disabled:lib:border-gray-200 peer-disabled:lib:text-gray-400",
+      invalid:
+        "lib:bg-red-600 lib:border-red-600 lib:text-white peer-hover:lib:bg-red-500 peer-hover:lib:border-red-500 peer-active:lib:bg-red-600 peer-active:lib:border-red-600 peer-focus-visible:lib:ring-2 peer-focus-visible:lib:ring-red-300 peer-disabled:lib:bg-red-200 peer-disabled:lib:border-red-200 peer-disabled:lib:text-red-400",
     },
   },
   text: {
-    base: "lib:text-gray-700 lib:select-none lib:font-medium",
+    base: "lib:text-gray-700 lib:font-medium lib:select-none",
     size: {
-      sm: "lib:text-xs",
-      md: "lib:text-sm lib:leading-4",
-      lg: "lib:text-sm lib:leading-5",
+      sm: "lib:text-cxs lib:ml-1.5",
+      md: "lib:text-sm lib:ml-2",
+      lg: "lib:text-base lib:ml-2",
     },
   },
-};
-
-export const checkbox = {
-  base: "group lib:relative lib:inline-flex lib:space-x-2 lib:items-center lib:align-top lib:cursor-pointer",
-  disabled: "lib:cursor-not-allowed",
-  input: "peer lib:sr-only",
-  icon: {
-    base: "lib:inline-flex lib:items-center lib:justify-center lib:flex-shrink-0 lib:align-top lib:self-start lib:select-none lib:transition-all lib:ring-0 lib:ring-gray-300 peer-focus:ring-2",
+  description: {
+    base: "lib:text-gray-500 lib:select-none",
     size: {
-      sm: "lib:w-3.5 lib:h-3.5 lib:text-[0.625rem] lib:rounded",
-      md: "lib:w-4 lib:h-4 lib:text-xs lib:rounded",
-      lg: "lib:w-5 lib:h-5 lib:text-base lib:rounded-md",
-    },
-    state: {
-      disabled: "lib:text-gray-400 lib:bg-gray-200 lib:cursor-not-allowed",
-      default: {
-        checked: "lib:text-white lib:bg-gray-800",
-        unchecked: "lib:border-[1.5px] lib:border-solid lib:border-gray-400",
-        indeterminate: "lib:text-white lib:bg-gray-800",
-      },
-      invalid: {
-        checked: "lib:bg-red-600 lib:text-white lib:ring-red-300",
-        unchecked:
-          "lib:border-[1.5px] lib:border-solid lib:border-red-600 lib:ring-red-300",
-        indeterminate: "lib:bg-red-600 lib:text-white lib:ring-red-300",
-      },
-      hover: {
-        checked: "group-hover:lib:bg-gray-700",
-        unchecked: "group-hover:lib:border-gray-500",
-        indeterminate: "group-hover:lib:bg-gray-700",
-      },
-      hover_invalid: {
-        checked: "group-hover:lib:bg-red-500",
-        unchecked: "group-hover:lib:border-red-500",
-        indeterminate: "group-hover:lib:bg-red-500",
-      },
+      sm: "lib:text-paragraph-cxs lib:mt-1 lib:ml-1.5",
+      md: "lib:text-paragraph-sm lib:mt-1.5 lib:ml-2",
+      lg: "lib:text-paragraph-sm lib:mt-2 lib:ml-2",
     },
   },
-  field,
 };
