+++ src/theme/defaultTheme/index.ts
@@ -5,7 +5,6 @@ import { button } from "./button";
 import { buttonGroup } from "./buttonGroup";
 import { calendar } from "./calendar";
 import { checkbox } from "./checkbox";
-import { checkboxNew } from "./checkboxNew";
 import { circularProgress } from "./circularProgress";
 import { datepicker } from "./datepicker";
 import { formField } from "./form-field";
@@ -32,7 +31,6 @@ const theme = {
   buttonGroup,
   calendar,
   checkbox,
-  checkboxNew,
   circularProgress,
   datepicker,
   formField,
